+++ RenderingEngine/timer.cpp
@@ -8,10 +8,10 @@ Timer::Timer()
 	lastFrameTime = high_resolution_clock::now();
 }
 
-double Timer::getDeltaTime()
+float Timer::getDeltaTime()
 {
 	high_resolution_clock::time_point currentFrameTime = high_resolution_clock::now();
-	duration<double, std::milli> frameTime = currentFrameTime - lastFrameTime;
+	duration<float, std::milli> frameTime = currentFrameTime - lastFrameTime;
 	return frameTime.count();
 }
 }
